Clean up the sbbs lib exported definitions.  Move all projects in subdirs of sbbs3 into the top-level list to avoid infinite recursion when a sub-projects include the sbbs3 list for libsbbs.so.

+++ xpdev/gen_defs.h
@@ -2,7 +2,7 @@
 
 /* General(ly useful) constant, macro, and type definitions */
 
-/* $Id: gen_defs.h,v 1.63 2014/02/06 08:19:02 deuce Exp $ */
+/* $Id: gen_defs.h,v 1.64 2014/02/06 20:37:32 deuce Exp $ */
 																			
 /****************************************************************************
  * @format.tab-size 4           (Plain Text/Source Code File Header)        *
@@ -176,8 +176,8 @@ typedef int64_t		intmax_t;
 #define _INTMAX_T_DECLARED
 
 #if !defined(HAS_STDINT_H)
-typedef long	intmax_t;
-typedef ulong	uintmax_t;
+typedef uintmax_t	uintptr_t;
+typedef intmax_t	intptr_t;
 #endif
 
 /* printf integer formatters: */
